requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.4

# Database
sqlalchemy>=2.0.0

# Slack Bot
slack-bolt>=1.18.0

# Utilities
python-dotenv>=1.0.0
